# web/htdocs/openapi/checkmk.yaml
components:
  parameters:
...
    ident:
      description: The identifier for this object. It's a 128bit uuid represented
        in hexadecimal (32 characters). There are no fixed parts or parts derived
        from the current hardware in this number.
      example: 49167bd012b44719a67956cf3ef7b3dd
      in: path
      name: ident
      required: true
      schema:
        pattern: '[a-fA-F0-9]{32}'
        type: string
...
  schemas:
    ApiError:
      properties:
        code:
          description: The HTTP status code.
          example: 404
          format: int32
          type: integer
        errors:
          description: Optionally a list of errors used for debugging.
          example: null
          items:
            type: string
          nullable: true
          type: array
        message:
          description: Detailed information on what exactly went wrong.
          example: The resource could not be found.
          type: string
        title:
          description: A summary of the problem.
          example: Not found
          type: string
      required:
      - code
      - message
      - title
      type: object
...
    DomainObjectCollection:
      properties:
        links:
          description: list of links to other resources.
          items:
            $ref: '#/components/schemas/Link'
          type: array
        value:
          items:
            $ref: '#/components/schemas/Link'
          type: array
      required:
      - links
      type: object
    Folder:
      properties:
        domainType:
          default: folder
          type: string
        links:
          description: list of links to other resources.
          items:
            $ref: '#/components/schemas/Link'
          type: array
        members:
          $ref: '#/components/schemas/ObjectMemberDict'
        title:
          type: string
      required:
      - domainType
      - links
      type: object
...
    InputAttribute:
      properties:
        key:
          type: string
        value:
          type: string
      required:
      - key
      - value
      type: object
...
    InputFolder:
      properties:
        attributes:
          example:
            foo: bar
          type: object
        name:
          description: The name of the folder.
          example: production
          type: string
        parent:
          description: The folder-id of the folder under which this folder shall be
            created. May be 'root' for the root-folder.
          example: root
          pattern: '[a-fA-F0-9]{32}|root'
          type: string
        title:
          example: Production Hosts
          type: string
      required:
      - name
      - parent
      - title
      type: object
...
    Link:
      properties:
        arguments:
          description: map that may be used as the basis for any data (arguments or
            properties) required to follow the link.
          nullable: true
          type: object
        href:
          description: The (absolute) address of the related resource. Any characters
            that are invalid in URLs must be URL encoded.
          example: https://.../api_resource
          type: string
        method:
          description: The HTTP method to use to traverse the link (get, post, put
            or delete)
          example: get
          pattern: get|put|post|delete
          type: string
        rel:
          description: Indicates the nature of the relationship of the related resource
            to the resource that generated this representation
          example: self
          type: string
        title:
          description: string that the consuming application may use to render the
            link without having to traverse the link in advance
          example: The object itself
          nullable: true
          type: string
        type:
          description: The media type that the linked resource will return
          example: application/json
          type: string
      required:
      - href
      - method
      - rel
      - type
      type: object
    ObjectActionMember:
      properties:
        disabledReason:
          description: Provides the reason (or the literal "disabled") why an object
            property or collection is un-modifiable, or, in the case of an action,
            unusable (and hence no links to mutate that member's state, or invoke
            the action, are provided).
          nullable: true
          type: string
        id:
          type: string
        invalidReason:
          description: Provides the reason (or the literal "invalid") why a proposed
            value for a property, collection or action argument is invalid. Appears
            within an argument representation 2.9 returned as a response.
          example: invalid
          nullable: true
          type: string
        links:
          description: list of links to other resources.
          items:
            $ref: '#/components/schemas/Link'
          type: array
        memberType:
          default: action
          type: string
        x-ro-invalidReason:
          description: Provides the reason why a SET OF proposed values for properties
            or arguments is invalid.
          nullable: true
          type: string
      required:
      - id
      - links
      type: object
    ObjectCollectionMember:
      properties:
        disabledReason:
          description: Provides the reason (or the literal "disabled") why an object
            property or collection is un-modifiable, or, in the case of an action,
            unusable (and hence no links to mutate that member's state, or invoke
            the action, are provided).
          nullable: true
          type: string
        id:
          type: string
        invalidReason:
          description: Provides the reason (or the literal "invalid") why a proposed
            value for a property, collection or action argument is invalid. Appears
            within an argument representation 2.9 returned as a response.
          example: invalid
          nullable: true
          type: string
        links:
          description: list of links to other resources.
          items:
            $ref: '#/components/schemas/Link'
          type: array
        memberType:
          default: collection
          type: string
        value:
          items:
            $ref: '#/components/schemas/Link'
          type: array
        x-ro-invalidReason:
          description: Provides the reason why a SET OF proposed values for properties
            or arguments is invalid.
          nullable: true
          type: string
      required:
      - id
      - links
      type: object
    ObjectMember:
      discriminator:
        mapping:
          action: '#/components/schemas/ObjectActionMember'
          collection: '#/components/schemas/ObjectCollectionMember'
          property: '#/components/schemas/ObjectPropertyMember'
        propertyName: memberType
      oneOf:
      - $ref: '#/components/schemas/ObjectActionMember'
      - $ref: '#/components/schemas/ObjectPropertyMember'
      - $ref: '#/components/schemas/ObjectCollectionMember'
      properties: {}
      type: object
    ObjectMemberDict:
      additionalProperties:
        $ref: '#/components/schemas/ObjectMember'
      properties: {}
      type: object
    ObjectPropertyMember:
      properties:
        disabledReason:
          description: Provides the reason (or the literal "disabled") why an object
            property or collection is un-modifiable, or, in the case of an action,
            unusable (and hence no links to mutate that member's state, or invoke
            the action, are provided).
          nullable: true
          type: string
        id:
          type: string
        invalidReason:
          description: Provides the reason (or the literal "invalid") why a proposed
            value for a property, collection or action argument is invalid. Appears
            within an argument representation 2.9 returned as a response.
          example: invalid
          nullable: true
          type: string
        links:
          description: list of links to other resources.
          items:
            $ref: '#/components/schemas/Link'
          type: array
        memberType:
          default: property
          type: string
        x-ro-invalidReason:
          description: Provides the reason why a SET OF proposed values for properties
            or arguments is invalid.
          nullable: true
          type: string
      required:
      - id
      - links
      type: object
...
    UpdateFolder:
      properties:
        attributes:
          example:
          - key: foo
            value: bar
          items:
            $ref: '#/components/schemas/InputAttribute'
          type: array
        title:
          example: Virtual Servers.
          type: string
      required:
      - title
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: username password
      description: 'The format of the header-value is "Bearer $automation_user $automation_user_password"


        Example: `Bearer hansdampf miezekatze123`'
      in: header
      scheme: bearer
      type: http
      x-bearerInfoFunc: cmk.gui.wsgi.auth.bearer_auth
...
info:
  contact:
    email: feedback@checkmk.com
    name: Contact the Checkmk Team
    url: https://checkmk.com/contact.php
  description: "API for Checkmk.\n\nThis is the docstring which will be the description\
    \ of the API.\n\n# Introduction\n\nThis API is documented in **OpenAPI format**.\
    \ This means there is a formal specification for the\nwhole API surface. This\
    \ documentation is generated from this specification.\n\n# Authentication\n\n\
    To use this API you have to create an *Automation User* via WATO. The resulting\
    \ username and\npassword have to be sent in the `Authentication` HTTP-Header in\
    \ the *Bearer* format. Example here:\n\n## Security definitions\n\n<SecurityDefinitions\
    \ />\n\n# How does it work?\n\nThis API follows the [HATEOAS](https://en.wikipedia.org/wiki/HATEOAS)\
    \ principle. While it can be\nused in a traditional way, using the elements that\
    \ enable HATEOAS make it way more powerful and\nflexible.\n\n## Text and payload\
    \ encoding\n\nAs specified by the `Content-Type` of `application/json`, the response\
    \ payload is serialized with\nJSON and encoded in UTF-8. Other formats are possible\
    \ in principle but currently not supported.\n\n## Links\n\n* Every response comes\
    \ with a collection of `links`. These links are used to provide the consumer\n\
    \  of the API with follow-up actions. For example a Folder response can have links\
...
  license:
...
    url: https://checkmk.com/gpl.html
...
openapi: 3.0.2
paths:
  /collections/folder:
    get:
      description: ''
      operationId: cmk.gui.plugins.openapi.endpoints.folder.list_folders
      responses:
        default:
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Any unsuccessful or unexpected result.
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainObjectCollection'
          description: ''
      summary: List folders
      tags:
      - cmk.gui.plugins.openapi.endpoints.folder
    post:
      description: This is the long description of the endpoint.
      operationId: cmk.gui.plugins.openapi.endpoints.folder.create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputFolder'
        required: true
      responses:
        default:
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Any unsuccessful or unexpected result.
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
          description: ''
          headers:
            ETag:
              description: The HTTP ETag header for this resource. It identifies the
                current state of the object and needs to be sent along in the "If-Match"
                request-header for subsequent modifications.
              schema:
                pattern: '[0-9a-fA-F]{32}'
                type: string
      summary: Create a new folder
      tags:
      - cmk.gui.plugins.openapi.endpoints.folder
      x-code-samples:
      - lang: curl
        source: "curl \\\n    -X POST \\\n    --header \"Authorization: Bearer username\
          \ password\" \\\n    --header \"Accept: application/json\" \\\n    -d '{\n\
          \          \"attributes\": {\n            \"foo\": \"bar\"\n          },\
          \ \n          \"name\": \"production\", \n          \"parent\": \"root\"\
          , \n          \"title\": \"Production Hosts\"\n        }' \\\n    /collections/folder"
      - lang: httpie
        source: "http POST /collections/folder \\\n    'Authorization: Bearer username\
          \ password' \\\n    attributes='{\"foo\": \"bar\"}' \\\n    name='production'\
          \ \\\n    parent='root' \\\n    title='Production Hosts'"
      - lang: requests
        source: "import requests\n\nresp = requests.post(\"/collections/folder\",\n\
          \    headers={'Authorization': 'Bearer username password'},\n    data={\n\
          \        \"attributes\": {\n            \"foo\": \"bar\"\n        }, \n\
          \        \"name\": \"production\", \n        \"parent\": \"root\", \n  \
          \      \"title\": \"Production Hosts\"\n    },\n)\nresp.raise_for_status()"
  /objects/folder/{ident}:
    delete:
      description: ''
      operationId: cmk.gui.plugins.openapi.endpoints.folder.delete
      parameters:
      - $ref: '#/components/parameters/ident'
      - description: The ETag of the object to be modified.
        in: header
        name: If-Match
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Any unsuccessful or unexpected result.
        '204':
          description: Operation done successfully. No further output.
      summary: Delete a folder
      tags:
      - cmk.gui.plugins.openapi.endpoints.folder
    get:
      description: ''
      operationId: cmk.gui.plugins.openapi.endpoints.folder.show_folder
      parameters:
      - $ref: '#/components/parameters/ident'
      responses:
        default:
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Any unsuccessful or unexpected result.
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
          description: ''
          headers:
            ETag:
              description: The HTTP ETag header for this resource. It identifies the
                current state of the object and needs to be sent along in the "If-Match"
                request-header for subsequent modifications.
              schema:
                pattern: '[0-9a-fA-F]{32}'
                type: string
      summary: Show a folder
      tags:
      - cmk.gui.plugins.openapi.endpoints.folder
    put:
      description: Title and attributes can be updated, but there is no checking of
        the attributes done.
      operationId: cmk.gui.plugins.openapi.endpoints.folder.update
      parameters:
      - $ref: '#/components/parameters/ident'
      - description: The ETag of the object to be modified.
        in: header
        name: If-Match
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFolder'
        required: true
      responses:
        default:
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Any unsuccessful or unexpected result.
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
          description: ''
          headers:
            ETag:
              description: The HTTP ETag header for this resource. It identifies the
                current state of the object and needs to be sent along in the "If-Match"
                request-header for subsequent modifications.
              schema:
                pattern: '[0-9a-fA-F]{32}'
                type: string
      summary: Update a folder
      tags:
      - cmk.gui.plugins.openapi.endpoints.folder
      x-code-samples:
      - lang: curl
        source: "curl \\\n    -X PUT \\\n    --header \"Authorization: Bearer username\
          \ password\" \\\n    --header \"Accept: application/json\" \\\n    -d '{\n\
          \          \"attributes\": [\n            {\n              \"value\": \"\
          bar\", \n              \"key\": \"foo\"\n            }\n          ], \n\
          \          \"title\": \"Virtual Servers.\"\n        }' \\\n    /objects/folder/49167bd012b44719a67956cf3ef7b3dd"
      - lang: httpie
        source: "http PUT /objects/folder/49167bd012b44719a67956cf3ef7b3dd \\\n  \
          \  'Authorization: Bearer username password' \\\n    attributes='[{'value':\
          \ 'bar', 'key': 'foo'}]' \\\n    title='Virtual Servers.'"
      - lang: requests
        source: "import requests\n\nresp = requests.put(\"/objects/folder/49167bd012b44719a67956cf3ef7b3dd\"\
          ,\n    headers={'Authorization': 'Bearer username password'},\n    data={\n\
          \        \"attributes\": [\n            {\n                \"key\": \"foo\"\
          , \n                \"value\": \"bar\"\n            }\n        ], \n   \
          \     \"title\": \"Virtual Servers.\"\n    },\n)\nresp.raise_for_status()"
  /objects/folder/{ident}/actions/move/invoke:
    post:
      description: ''
      operationId: cmk.gui.plugins.openapi.endpoints.folder.move
      parameters:
      - $ref: '#/components/parameters/ident'
      - description: The ETag of the object to be modified.
        in: header
        name: If-Match
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Any unsuccessful or unexpected result.
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
          description: ''
          headers:
            ETag:
              description: The HTTP ETag header for this resource. It identifies the
                current state of the object and needs to be sent along in the "If-Match"
                request-header for subsequent modifications.
              schema:
                pattern: '[0-9a-fA-F]{32}'
                type: string
      summary: Move a folder
      tags:
      - cmk.gui.plugins.openapi.endpoints.folder
...
security:
- BearerAuth: []
...